[REF] *: remove isCheck in tours
As of commit https://github.com/odoo/odoo/pull/167743, the default
action in tours is now NOTHING. That is to say that if we only put
a "trigger" in a step, the compiler only checks that the element is
in the DOM.
To perform a "click" action on the trigger, you must explicitly write
run: "click".
In this commit, we remove all isCheck: true and empty runs because
they are no longer useful... since it is the default action.

task~3974087

closes odoo/enterprise#64080

Related: odoo/odoo#168474
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>