[IMP] spreadsheet_edition: Allow to rolback history to first state
Currently, a user can only roll back on revisions and not on the
original state of the spreadsheet (i.e. the original json in the
filestore).
This means that for users that upload an xlsx file and start by doing an
huge modification on it, they don't have the possibility to return to
the spreasheet as it was uploaded and had no choice but to reupload it.

This revision allows the user to roll back to the initial state of the
spreadsheet.

task-3799678

closes odoo/enterprise#63974

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>