[FIX] appointment: do not create 2 slot of 15 min
When trying to create a custom slot of 15min in the calendar,
two slots were created.
This was because the fullcalendar library was detecting two events:
a date selection and a date click.

The date selection is correct but the size of the event (15 min) is
the minimal size which execute a click using the default time.

To prevent this behaviour, we use preventDefault to ignore the second
event and so the creation of the second slot.
A similar fix was done in calendar which was ignored because of the
slot-creation mode.
See related commit at odoo/odoo@38331cbaaaf0a700493108c238eacc70bc4f135a

task-3976122

closes odoo/enterprise#65366

X-original-commit: c8f7461110d56dd9a2ce2a5d6383ad048265089c
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>