[FIX] account_bank_statement_import_camt: fix payment_ref priorities
In previous commit https://github.com/odoo/enterprise/commit/099531438fe9ef7b008a2b281780ed3ee2175692
we added a new tag check. As we tought it was empty for non-nordic countries, we thought
it was fine to have it in first place but happens it is sometimes filled in other countries
but not with the payment reference, as in Switzerland.

We now check to fill payment_ref, in order:
1. NtryDtls>RmtInf>Ustrd (same as before changes)
2. NtryDtls>RmtInf>Strd>CdtrRefInf>Ref (same as before changes)
3. Ntry>AddtlNtryInf (was buggy before https://github.com/odoo/enterprise/pull/62703)
4. NtryDtls>RmtInf>Strd>AddtlRmtInf (lowered priority)

opw-3923665
opw-3971080
opw-3932651
opw-3918491

closes odoo/enterprise#64220

X-original-commit: a2003125e31fa71ad671c8af946b9f9cc5b07ae3
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>