[FIX] sale_subscription: avoid making payment with negative amount
Before, when the subscription cron create recurring invoices, it might create
a credit note in case some already invoiced non recurring product had his
ordered quantity set back to zero.
The payment transaction created for this credit note will take the absolute
total without taking into account the type of the entry process. This leads
to a transaction charging the amount to the client instead of refunding him.

As we don't want to allow automatic refund because it could lead to
catastrophe if a bug occurs, we decide to prevent the payment request if
the amount is negative. In that case, we keep the draft invoice and keep
the "payment_exception" flag to avoid running the subscription invoice
cron again on that contract.

opw-3874964

closes odoo/enterprise#64098

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>