[IMP] knowledge: define calendar start/end display
This commit adds new editing options to embedded calendar views.

Sometimes a user may want to limit the scope of its calendar to a reduced time
interval (i.e. only diplay hours from 09h00 to 17h00 in day scale) to avoid
displaying a blank space.
To do that, new inputs are introduced in the ItemCalendarPropsDialog which
enables the user to select a starting and ending day for a week, and a starting
and ending hour for each day.

The new inputs are affecting the following options for the calendar:
 * `slotMinTime`: determines the hour at which the first slot starts
 * `slotMaxTime`: determines the hour at which the last slot ends
 * `weekends`: determines if we show the weekend

task-3902124

Part-of: odoo/enterprise#64006
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>