[FIX] stock_barcode: ensure req packaging in cache
**Current behavior:**
A picking with a move for a product with kit BOM that also has
some product packaging can be un-openable in Barcode.

**Expected behavior:**
Any picking should be able to be processed in Barcode.

**Steps to reproduce:**
*Enable product packagings*

1. Create a storable product with a kit BOM and a product
     packaging

2. Make some picking that has a move for the new product,
     confirm it so the moves for the kit components are created

3. Add stock move lines for each of the moves, fill out their
     `qty_done` field with the required amount

4. Save the picking record, go into Barcode

5. Try to open the picking -> trace back

**Cause of the issue:**
The Barcode client tries to read the package record of the BOM
product/product template but it was never fetched into the
barcode cache.

**Fix:**
Ensure any packaging record that could have an attempted access
is fetched by the barcode client `onWillStart()`.

opw-3895426

closes odoo/enterprise#65179

X-original-commit: 6a5aad926eb5a0cf668453613b4765b04c4e69ea
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Vincent Ethan <etvi@odoo.com>