[FIX] l10n_mx_edi: change fecha date to current date
Issue: Sometimes payments are created in Odoo prior to sending the CFDI, by law payments don't have to be necessarily sent
to the government on the same month and can sometimes be sent up to the first 10 days of the next month.

Solution: fill the 'fecha' field with the current datetime for the cfdi of the payments, Then send it.

Task-3885769

[IMP] l10n_mx_edi*: Add timezone tests for 'Fecha' computation

This task has been merged: 819ed2a
...then reverted: 3f113f1

This is not the first time we create issues due to timezone in Mexico.
Let's add tests for that to prevent issue in the future.

Forward-port-of: https://github.com/odoo/enterprise/pull/63488
X-original-commit: ccbd30c675faec3116e4042eb8ed96655a507545
Part-of: odoo/enterprise#65259
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>