[FIX] l10n_mx_edi: change fecha date to current date
Issue: Sometimes payments are created in Odoo prior to sending the CFDI, by law payments don't have to be necessarily sent
to the government on the same month and can sometimes be sent up to the first 10 days of the next month.

Solution: fill the 'fecha' field with the current datetime for the cfdi of the payments, Then send it.

Task-3885769

closes odoo/enterprise#63488

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>