[FIX] appointment: restore expected search domain behavior
Following odoo/odoo@0bc65a8 ,
the empty domain [] is normalized as True, which means that
used with expression.OR, as it is the case in appointment portal,
it will accept all records, leading to a useless search.

Instead, append appropriate domains and use OR on the whole list,
before using AND on the default domain, to restrict records to
the search, as expected.

Task-3976616

closes odoo/enterprise#64276

X-original-commit: 516e211176592e30955ce1027c87fc3debc741de
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>