[FIX] web_studio: translate studio approval message
Problem: When the user enable approval rules for confirming
purchase orders, the approval message was not being translated
to the user's preferred languages.

Purpose: Since multiple users can be notified for the approval,
it was discussed that translating the message to be in the
language of the first user should help with the limitation.
It will solve the bug in the case only one user needs to be
notified.

Steps to reproduce on Runbot:
1. Install Studio and purchase
2. create a purchase order without confirming it
3. Go into Studio and edit the "confirm order' button on the PO
4. Add 2 approval rules
- notify User A with notification order 1
- notify User B whose language is not English with notification order 2
5. Confirm the PO and observe the message is sent in English

opw-3932725

closes odoo/enterprise#64060

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>