[FIX] knowledge: add overflow to calendar year view
This commit adds a scrollbar to the calendar embedded view for the year view,
which isn't present.
This scrollbar should've been present as the complete year of the calendar
cannot be showed in its fullness otherwise.

To fix this, the overflow-hidden bootstrap class is now ignored via a CSS rule
that unsets its overflow value. This is applied only for embedded calendar
views, as the regular view doesn't have any issue.

task-3902124

Part-of: odoo/enterprise#64006
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>