[FIX] appointment: fix multi company context
Before, in a multi company and mutli website setup,
booking an appointment used the companies on the
organizer for the context.
This was leading to issue for appointment type in
Website 2 tied to Company 2. Indeed the mails for the
booking were sent with the context of Company 1 (value
based on the allowed_company_ids of the user).

Now, when website_appointment is installed, if a website
is linked to the actual request we use the company tied to it.
This ensure that we always use the correct company of the website
for the creation of the event.

In case of appointment type with resource, where we don't have
staff user, we take the create_uid of the appointment type as
a fallback like we do with the organizer of the event.

We target 17.0 to avoid a behavior change in older stable as the
multi website was officially supported for appointment type in
that version. Could be backported if really necessary.

task-3977787

closes odoo/enterprise#67420

X-original-commit: 4c7ef52b048a072dce244e47554f1fd2fac87c4c
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>