[IMP] knowledge: redirect internal users to new routes
When the user accesses an article via a shared link (`/knowledge/article/42`),
the internal users will be redirected to the backend view of Knowledge
using the legacy route (i.e: `/web#id=...&model=...`). When loading the
page, the router will transform that url to a more friendly one
(i.e: `/odoo/knowledge/...`). To avoid updating the url many times, we
will directly redirect the user to the new route.

task-3975577

closes odoo/enterprise#68331

X-original-commit: c86b09053fccf6f56ef4060930f763c1e3ace579
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>