[FIX] account_reports: make the (No Group) lines fold individually
In the reports the groups called '(No Group)' fold and unfold together.

Steps to reproduce in runbot:
- in Accounting app go in Configuration > Account Groups and create an account group for the current company
- Go in Reporting > Profit and Loss
- All '(No Group)' lines fold and unfold together

Cause:
Line ids are duplicate for (No Group) lines: '~account.group~'. The hierarchy is ignored for these ids, there are no reference to the parent line.

Solution:
Integrate the parent line in the ids to take the hierarchy into consideration.

opw-3885400

closes odoo/enterprise#64219

X-original-commit: fd7c3fa377adee53a4177a24997a328be5cebcea
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>