[FIX] sale_subscription: save token on SO when paying w/ Adyen
A new transparent way to save payment methods on a subscription has been
introduced in 36b238b934f1b7e4ddf1dfd6089d36ea0b8ac7eb. This consists of a checkbox "Automate payments
for this subscription" that is shown in the payment method form when
paying invoices. When this checkbox is checked, an override of
`_initiatePaymentFlow` changes the transaction route to Subscription's
to ensure that the token is saved on the sale order. Adyen
implementation doesn't call the parent `_initiatePaymentFlow` leading to
the token not being stored on the sale order.

To avoid such cases in the future, the override of `_initiatePaymentFlow`
has been replaced by an override of `_submitForm`. Now, when the
checkbox is checked, the token will be saved on the sale order.

opw-3893592
opw-3946899

X-original-commit: e16a3b39de5168e29b8141dfaadda1364982bd34
Part-of: odoo/enterprise#64419
Signed-off-by: Valentin Chevalier <vcr@odoo.com>