[FIX] sale_subscription: select the right checkbox
A new transparent way to save payment methods on a subscription has been
introduced in 36b238b934f1b7e4ddf1dfd6089d36ea0b8ac7eb. This consists of a checkbox "Automate payments
for this subscription" that is shown in the payment method form when
paying invoices. The checkbox of the automated payment sought may be
that of another payment method.

Now, the checkbox will be sought based on the selected payment method.

opw-3893592
opw-3946899

X-original-commit: 9d7b51aff961efb6c0414fcb23e63e73f290e7bc
Part-of: odoo/enterprise#64419
Signed-off-by: Valentin Chevalier <vcr@odoo.com>