[FIX] sale_subscription: show sub checkbox in payment tokens
Commit 36b238b934f1b7e4ddf1dfd6089d36ea0b8ac7eb introduced a new transparent way to save payment methods
on a subscription. This consists of a checkbox "Automate payments for
this subscription" that should be shown in the payment token inline form
when paying invoices. Currently, it's only shown for tokens of two
providers Demo and Razorpay.

Now, the checkbox will be shown whether the token has an inline form or
not.

opw-3893592
opw-3946899

closes odoo/enterprise#64419

X-original-commit: c7b44f7976b413f056f9d50c41a7f41e29e9d582
Signed-off-by: Valentin Chevalier <vcr@odoo.com>