[FIX] sale_stock_renting: only show rented lots
**Current behavior:**
A product that is tracked via serial will have all of its lots
displayed in the Schedule gantt view instead of just the ones
which are part of the order.

**Expected behavior:**
Only the lots corresponding to rented quants of a product will
be displayed.

**Steps to reproduce:**
1. Enable `Rental Transfers` in settings

2. Create a rental order with 1 of the demo printer products
     (they are tracked via SN already) and confirm

3. Open the rental schedule view, see that for the new order,
     all of the printers in stock with a SN are displayed

**Cause of the issue:**
Previously, the commit 8c90c7493ec was merged to fix an issue
where you could not group by serial numbers in the schedule view
for orders that were created while the 'Rental Transfers'
setting was enabled.

This introducted this bug by not actually specifying which
`lot_id`s of a product were part of an order- so it selected ALL
`lot_id`s of ANY product in a rental order.

**Fix:**
Add a join on `StockMove` so we can select `lot_id`s which are
exclusively in `StockMoveLine`s that are linked to `StockMoves`
which correspond to a rental order.

opw-3977685

closes odoo/enterprise#64118

Signed-off-by: Stéphane Diez (snd) <snd@odoo.com>