[FIX] documents_spreadsheet: Fix Odoo graph data name
Since the recent chart options refactoring, the title of the chart is
not longer a string but an object with different style-related keys and
the title string is set in the key `text`. The getter
`getOdooChartDIsplayName` was not adapted to this change.

This commit adapts the tests in Enterprise.

task-3978637

closes odoo/enterprise#64392

X-original-commit: 09be9298226f3243a4979468ca8a9b767ae4e89b
Related: odoo/odoo#168965
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>