[IMP] web_studio: improve fieldsToChoices util
This commit improves the usage and ability of this utility function to filter
appropriate fields to display in the Sidebar UI. Many usages of this function
verified the type of the fields globally available. Now, it is easier to list
the desired types, and the additional filtering callback can still be given,
when filtering even more the available fields to show.

Part-of: odoo/enterprise#64296
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>