[IMP] web_studio: display the name of the structure in the sidebar
This commit improves the user experience by displaying the name of the selected
element in the sidebar, and display a more human name in the deletion dialog.

This is handful when working in elements being boxed in others:
e.g. a kanban card, containing many divs for the layout, near to each other

Part-of: odoo/enterprise#64852
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>