[IMP] web_studio: allow edition of 'class' attribute
Before this commit, it was not possible to change the value of the class attribute,
except for the button element (which added a dedicated tooltip). This commit brings
this ability to the editor, directly from the sidebar, and adds a common tooltip,
regarding the element being selected.

A test has been added to verify this could be achieved, and that edit_view has the right
value. QUnit tests that needed to be modified has been converted to HOOT directly as well,
and the legacy tests have been removed.

closes odoo/enterprise#64296

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>