[FIX] account_accountant,accountant: fix security groups
The `basic` accounting group was added to differentiate between invoicing enterprise and regular invoicing.
It grants access to the bank recon and other basic accounting features, and is a hidden group.

Problem
-------
Since the basic group is hidden, it is not removed when downgrading a user to a group that does not imply the basic group.
Therefore `Billing` users keep access to basic functions.

Solution
--------
Add the basic group to the Accounting category.
Make Bookkeeper group visible again

closes odoo/enterprise#64635

X-original-commit: 8d39979dbe4d9ee3331b74d5f2d0883c841cb68c
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>