[FIX] pos_restaurant_appointment: show delays
In `pos_restaurant`, the user will see the upcoming appointment for each
table underneath it. The text should be  styled differently based on
the situation. Late arrivals should be shown in red; future appointments
should be shown with decreased opacity.

During the refactoring in bb06fde, a bug was introduced  that made the
time computation invalid. This commit addresses the issue.

closes odoo/enterprise#64343

Task: 3978801
X-original-commit: ce4cfc1a953e7616b5f3bcd76c5a01c00463c68f
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>