[FIX] appointment: missing cancellation notification
Steps to reproduce
==================
​1).Log as Mitchell, create an appointment and set Joel as an attendee
2).Log as Joel, open the appointment page, and cancel
-> cancellation mail is not send

Before this commit
==================
A cancellation email is only sent when the person who books the appointment
cancels the appointment.

Technical
==========
https://github.com/odoo/enterprise/pull/45332/commits/46f1b8aaeaa94d3ac7b72784a8fa7120f8a3e8a5#diff-394903898ffdd9f9e1f053ecda2afd00cda86d6cc96159268fbe4fefc62f3a83R187
Before the above commit we were archiving the meeting if there are < 2
attendees in the event.
So now we restore that condition with the current conditions

After this commit
=================
Now Cancellation email will also be sent if an event contains less than 2
attendee after the cancellation.

Task-3975493

closes odoo/enterprise#64238

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>