[FIX] stock_barcode: don't display not scanned lot
Before this commit, the reserved lot for a line was displayed even if
the "Show reserved lots/SN" is not checked.

How to reproduce:
- Ensure "Show reserved lots/SN" is not check for the deliveries;
- Create a product tracked by lots and set a barcodes;
- Create two lots for this product and add quantity in stock for both;
- Create a delivery for this product (enough qty. to reserved both
  lots);
- Open the delivery in the Barcode app.

Issue:
On the barcode line, the first reserved lot is displayed.

Expected behavior:
In this configuration, only scanned lots should be displayed (here, the
barcode line lot part should appear empty.)

This commit also fixes an issue where not the same move lines is opened
when using the edit form view in the Barcode App.
To reproduce, follow steps above, then open the display line in the form
view (via the edit button) => You can see a lot is displayed.
Update the quantity to fulfill the line, save the change and then
re-open the form view => You can now see the other lot and 0 qty because
it's the other reserved move line who was opened.

To avoid this issue, it will always be the first (by ID) move line who
will be opened.

closes odoo/enterprise#66249

X-original-commit: a8afa1d1b769a4c0a5e4316453fdd3137f1d06c1
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>