[FIX] voip: no crash on adding removed call activity
When call activity gets deleted or added, the related component has to
be re-rendered so that most uptodate information is shown, otherwise
this might lead to crash if those activities are accessed. To mitigate
the issue, component re-rendering is invoked when call activity gets
added or removed from the queue.
The activity removal happens when `delete_call_activity` function is
invoked directly, which crafts domain, to search the activies that are outdated.

Steps to reproduce:

- on crm lead schedule call activity and make sure that activity is old
- book activity and save it
- add it to Call Queue
- remove it from Call Queue
- try to access it/click on it

X-original-commit: 2ed3780ade29b0e1f404aca85dccf11f1dbb21a1

.

closes odoo/enterprise#64476

X-original-commit: edb5ea7c5ba8d80d5746aabd12e99f571e820c5b
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>
Signed-off-by: Dato Nefaridze (dane) <dane@odoo.com>