[FIX] account_invoice_extract_*: fix undeterminism in tests
The test `test_action_reload_ai_data` added by commit f85e392 introduced
a new undeterministic failure when the test is run at around midnight.

This is the exact same issue that occurred in another of these tests and
that was fixed by commit a4fbd4c.

This commit solves the problem in a more generic way, by freezing the
date so that the `create_date` and the default `invoice_date_due` always
match. It should prevent the issue from re-appearing in future new
tests.

closes odoo/enterprise#64569

X-original-commit: 2a453bd4bfe286c485307d2cbed7e96fee3fe7c2
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>