[FIX] timesheet_grid: only remind employees in right company
Duplicate emails were being sent out to employees because employees were
being searched from user ids for each company, but each search was not
filtered by the company. So if there was a user that had multiple
employee records, all of their employee records on every company would
be pulled up for every company on their database that had timesheet
notifications on. For example if there were 4 companies and the employee
had an employee record for each of them they would receive 16 emails in
total. 4 for each employee record.

Adding to the search domain to only look for employees on the current
company defined by the for loop avoids these duplicate emails. This is
similar to a fix that was done on the manager notification emails here:
https://github.com/odoo/enterprise/pull/52808

opw-3748426

closes odoo/enterprise#65699

X-original-commit: 21213479828eb280e0ce85d35a5994b5a81f7d0a
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Ryan Cen (ryce) <ryce@odoo.com>