[REF] web_studio: add ViewStructures component
The code is moved into a component dedicated to display some common structures
in a view (columns and notebook). This will allow to use the same kind of separation
in the sidebar for structures in other editors too, such as kanban.

A humanName can be set on structures to make use of it.

Part-of: odoo/enterprise#64296
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>