[REF] web_studio: create new api kanban editor
This commit creates a new kanban editor for the new arch API. The legacy kanban editor
using the old one is still active for compatibility purposes until we stop the support
of those in upcoming versions of Odoo.

The cleanup of that old Kanban editor (using the deprecated kanban-card API) will be easy.
For example, changes made to the ViewEditorModel can simply be undone, and the
kanban_legacy folder deleted, once the support is dropped.

This new editor makes it easy to add more features to the Kanban view than ever before, since
we can adjust the layout, using predefined structures, available directly from the sidebar.
Then, it is possible to drag and drop any field inside those structures.

Some widgets can be added directly inside the kanban card (kanban_image or web_ribbon).

Tests and tours have been added or adapted to verify the correct behavior when using this
new editor.

part of task-3650117

closes odoo/enterprise#65593

Related: odoo/odoo#170986
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Co-authored-by: Florent Dardenne <dafl@odoo.com>