[FIX] l10n_be_reports: fix partner VAT listing warnings
Currently the warning of the partner VAT listing is never shown.

In 88b449b08e21700e9dd113250125b2c3daa65a0f (17.2) the warnings for custom
report handlers were moved from `_custom_line_postprocessor`
to a new function called `_customize_warnings`.
Currently this is not the case for the partner VAT listing.

This commit does the necessary change for the partner VAT listing.

(This commit does not belong to any task)

closes odoo/enterprise#66277

X-original-commit: 707fa4b5298d390f7419658e3d663d88abf8bdbc
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>