[IMP] account_accountant{_batch_payment}: improve bank reco widget perf
When adding a huge batch of payments, the bank reconciliation widget
gets very slow for every action taken. It is due to the fact that adding
a batch loads every payments line. Any following action will then apply
the onchange for each of these lines (hence the slow down).

The goal here is to allow mounting a batch payment as one line in the
bank reconciliation widget and push all possible computing at the
validation step.

task-3607121

closes odoo/enterprise#63467

Signed-off-by: Laurent Smet (las) <las@odoo.com>