[FIX] stock_barcode: always do backorder dialog
**Current behavior:**
Having a picking whose operation type has the
`create_backorder='ask'` setting, if a barcode line is modified
using the edit button rather than actually scanning, the
confirmation dialog for the backorder will be skipped.

**Expected behavior:**
When the 'ask' option is used, the dialog should always be
displayed.

**Steps to reproduce:**
1. Create a picking with a demand of 2.0 for some product,
     ensure the picking's op type has the 'ask' options set for
     backorder creation (it's the default option) -> confirm the
     picking

2. In Barcode, open the picking and add 1.0 quantity to the
     barcode line

3. Save the edit form, then validate the picking -> no dialog

**Cause of the issue:**
Compared the actually scanning product, modifying the line via
the form actually updates the record data. This causes the
corresponding `StockMoveLine` to get marked as `picked=True`
(via `_inverse_qty_done()` in `stock_barcode::StockMoveLine`.

The end result is `!this.lineIsReserved(line)` evaluates to
True in the Barcode client when checking whether to do
the backorder dialog which means we skip it.

**Fix:**
Remove the `lineIsReserved()` check, only look at whether
the line quantity is positive, non-zero.

*Note: the behavior between the two flows (scans vs. form edits)
diverges when the user reloads the page, however, this was also
the case prior to this change.*

opw-3972478

closes odoo/enterprise#66714

X-original-commit: e2222ac1ecae1c78b5ac3bfabce973b3d6381439
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Vincent Ethan <etvi@odoo.com>