[FIX] documents_spreadsheet: prevent edition of revision
Revisions are technical records that are never supposed to be changed
(except the `active` field). If any of `res_model`, `res_id`, `revision_uuid`
is changed, the spreadsheet history will be broken.

Changing field `commands` could also break the spreadsheet if the json is
malformed or the command is made invalid.

In this commit, we prevent the admin to make any changes.

closes odoo/enterprise#64779

X-original-commit: 887a617893bd3ecee2d4f533cd344f8c6bcc0a28
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>