[FIX] documents_spreadsheet: prevent edition of revision
Revisions are technical records that are never supposed to be changed
(except the `active` field). If any of `res_model`, `res_id`, `revision_uuid`
is changed, the spreadsheet history will be broken.

Changing field `commands` could also break the spreadsheet if the json is
malformed or the command is made invalid.

In this commit, we prevent the admin to make any changes.

closes odoo/enterprise#64328

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>