[FIX] sign: selection item options not working properly
Before this commit, selection item options were not updated properly.

1. When adding some options, validating and then removing the options
and go to sign, the removed options would still be there.

This happened because when passing an empty array to the backend
it would be recognized as an empty set of commands,
it's like we are telling odoo to do nothing.

To fix this we use the command [6, false, [new_ids]].

2. `state.option_ids` were updated using `changes.option_ids` which is very fragile,
for example in 16.0 we the changes are given in the form of `[6, false, [id1, id2, ..]]`
but in 17.0 we have a list of commands for adding and removing for example `[[4, id1], [3, id2]]`
To count on some fixed form of updates would introudce bugs if this form changes in future.

To fix this we use `record.data.option_ids.currentIds` which is basically the live set of Ids.

3. when updating both options and responsible at the same time before validating,
responsible won't update. This happened becase of the return statement when the options change.

To fix this wer remove the return keyword and allow them both to be updated.

closes odoo/enterprise#66026

Task: 3972994
X-original-commit: 0278838f6e42fea18b9d815c080feeeea7e5506a
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>