[FIX] l10n_nl_reports: Skip forbidden country check if no XSD present
In https://github.com/odoo/odoo/pull/115720 downloading of XSD's for
validation was made optional and related checks were skipped if the
relevant XSD was not present. That's currently not done when generating
a XAF report in the Dutch localization. The XSD contains the list of
valid countries allowed in the report. If there is no XSD, no country
is allowed.

This can result in the user being presented with an error that partners
in the Netherlands are not allowed.

This fix skips the country check if the XSD is not present. The downside
of that is not having this validation by default, since the XSD is not
downloaded by default. An alternative would be to hardcode the list.

opw-3957513

X-original-commit: c3cabd29c6aad81373fb372e2685d563aac0f810
Part-of: odoo/enterprise#64563
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>