[IMP] website_helpdesk: only log message in chatter
Before this commit, we got custom code to add custom message and add
additional message to the description of the ticket.

This commit just log the message if the default field set on the website
is the description field of the ticket field since now the description
field will already contain the additional information gathered by the
formulae filled by a public user.

task-3419255

closes odoo/enterprise#64352

Related: odoo/odoo#128964
Related: odoo/upgrade#5436
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>