[FIX] l10n_nl_reports: Missing prefix when checking for XSD
When generating a XAF report, allowed country codes for partners are
loaded from an XSD. By default this XSD is not present. If it's manually
downloaded through the settings, it still didn't work because the name
of the attachment searched for does not include the module prefix while
the name of the downloaded attachment does.

This fix includes the prefix in the search.

opw-3957513

closes odoo/enterprise#64563

X-original-commit: 124dd72c5df57873edd66f1d719da5947f0b0e62
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>