[FIX] pos_iot: customer screen using iot-connected display
When selecting the option "IoT-connected display" for customer display, and
pos_iot is installed, there is no way from the res.config.settings form to
specify the `proxy_ip` field which is used to send the customer display
information. Because of this, using IoT-connected display for customer screen
doesn't work.

This commit introduces a way to infer the proxy_ip from the specify display
device.

closes odoo/enterprise#64612

X-original-commit: b58359f7cfca534bdfad932029256bfc3fe0f474
Related: odoo/odoo#169286
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>