[PERF] account_accountant: add CTE to invoice amls candidates query
When triggering the onchange method of the reconciliation widget,
the `_get_invoice_matching_amls_candidates` method is executed. If there
are both numerical and exact tokens, it produces a query with 6 subqueries
(a general query with 6 UNION ALL). This can become quite slow
on database with lots of amls.

Because for each of these 6 "subqueries" the FROM, JOIN and WHERE clauses
are the same, we can extract all of them in a single CTE. Thanks to that
the CTE is executed only once and each UNION ALL simply scans the CTE
to get the correct columns + performs an additional filtering. This
reduces the average execution of the query by a factor of 3-4 on hot-cache.
Since the onchange is triggered each time we do something in the
reconciliation widget, on average the cache is expected to be hot so
the average speedup boils down to the hot-cache scenario.

speedup

Customer 17.0 database with 7M account_move_lines and 8000 accounts,
the query execution, given a st_line and with both numerical and exact
tokens: 7.5s -> 2.3s

opw-3938995

closes odoo/enterprise#64730

X-original-commit: eff1b417b16044cc68ae17ef6e74740f489e2314
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Aurélien van Delft (avd) <avd@odoo.com>