[FIX] timesheet_grid: fail TestOvertime without hr_contract
Before this commit, the test fails when `hr_contract` module is not
installed because the test tries to create a contract for an employee.

This commit makes sure the test works without `hr_contract` by saying
the employee is a freelancer to avoid using contract to get his working
schedule to know if he did an overtime or not in Timesheets app.

closes odoo/enterprise#64555

X-original-commit: 1e58628c63c4951dfb66ec1aa3a7eeae0a7b78d4
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>