[IMP] social_*: customize the social post based on the media
Purpose
=======
Allow customizing the content of the social post (the message and the images),
based on the media on which we are posting.

Specification
=============
Allow to split the configuration for all selected medias, each selected medias
have one tab in which we can change the message or the images and see the preview.

Each medias has its own message and images field. Adapt the form view for that new
feature (the general preview is only visible if we don't split per network,
add a button to split, once it's splitted we can not go back).

Misc
====
Show the YouTube thumbnail in the social posts kanban view only if
there's no image on other medias.

Task-3109395

closes odoo/enterprise#64577

Related: odoo/upgrade#6157
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>