[IMP] account_accountant: new "full_months" computation method for deferred
The computation methods have been renamed:
- Based on days -> Days
- Equal per Month -> Months & First Day
- Months (new full months option)

In this new computation method, when a month is started, we fully account
a "equal per month" amount for it, instead of a prorata-ed amount. The last
month will null to compensate for the full first month.

E.g.:

1. 12000 from 2024-06-05 to 2025-06-04:
- 1000 for June 2024 -> Dec 2024, and Jan 2025 -> May 2025
- 0 for June 2025

2. 12000 from 2024-06-30 to 2025-06-29:
- 1000 for June 2024 -> Dec 2024, and Jan 2025 -> May 2025
- 0 for June 2025

3. 12000 from 2024-07-01 -> 2025-06-30:
- 1000 for July 2024 -> Dec 2024, and Jan 2025 -> June 2025
- 0 for July 2025
- same behaviour as "Months & First Day" because the months are full

More examples in the tests

task-id 3925925

closes odoo/enterprise#62470

Signed-off-by: William André (wan) <wan@odoo.com>