[FIX] website(_helpdesk)_knowledge: exclude templates from search
When we introduced dynamic templates in Knowledge, we merged the models
`knowledge.article` and `knowledge.article.template` and introduced a
new field `is_template` in the `knowledge.article` model to distinguish
the templates from the regular articles.

In the `website_helpdesk_knowledge` module, we forgot to exclude the
templates from the search domains. As a result, the helpdesk frontend
view can list and search within templates.

Steps to reproduce the issue:

1. Log in as "Admin".
2. Install `website_helpdesk_knowledge` module.
3. Go on the Helpdesk module.
4. On the helpdesk overview, click on the vertical dots of the
   "Customer Care" team card.
5. On the dropdown menu, click on the "Settings" option.
6. On the "Help Center" section, click on the "Knowledge" option.
7. On the "Visibility & Assignment" section, click on "Invited portal
   users and all internal users" option.
8. Save the changes.
9. Go on the Website module.
10. On the website navbar, click on "Customer Care" nav bar.
11. The interface should list article templates.
=> TO BE: The system should not list any templates.

To fix that issue, we will simplify the search domains and exclude the
templates from the search results using the `is_template` field.

Reference: https://github.com/odoo/enterprise/pull/48691

task-3987469

closes odoo/enterprise#65351

X-original-commit: 671f5343f0c0faf200c5333b67c71737c69d7701
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>