[FIX] account_reports: enhance journal report pdf readability
In accounting, when exporting journal report to pdf, the Taxes Applied table size is limited thus forcing long number to wrap.

Steps to reproduce:
1.go to Accounting > Reporting > Audit Reports > Journal Report
2.click on Pdf button
3.the "Taxes Applied" table will be spanning accross 2 column of the parent table leading to wrapped values

Cause:
"Taxes Applied" and "Tax Grid" are sharing the same row to render 2 different table and there is not enough room
especially if the numbers are long.

Solution:
Up the colspan for "Taxes Applied" to 3, not the ideal solution if each column are diplayed and filled with long numbers but it is stable friendly.

opw-3959560

closes odoo/enterprise#64467

Signed-off-by: William André (wan) <wan@odoo.com>