[FIX] frontdesk: make sure visitor_id is set when ordering drink
If the network is really slow, when you order a drink at the frontdesk.
The visitor_id will not be set correctly, and you will have a traceback.

Steps to reproduce:
-------------------
* Install frontdesk
* Enable "host selection" on the frontdesk station
* Choose open kiosk
* With the developper tool make the network slow, and use a mobile view
* Click on check in, then try to order a drink

> Observation: You get a traceback

Why the fix:
------------
The `createVisitor` function was not awaited during the setup of the
page. But it is required to order the drink.

opw-3889245

closes odoo/enterprise#62899

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>